refactor(types): rename tool-type file and update its import
