+++ src/data/tools.ts
@@ -1,10 +1,9 @@
-import { Tool } from '@/types/tool-type';
+import { Tool } from '@/types';
 
 export const tools: Tool[] = [
   {
     name: 'URL Shortener',
-    description:
-      'Link shortening service with free-of-charge advanced management features.',
+    description: 'Link shortening service with free-of-charge advanced management features.',
     iconFilenames: ['shortener.svg', 'shortener_1.svg'],
     secondaryIconClass: 'right-[-8%] top-[20%] w-[50%]',
     decorImgSrc: '/images/home/side_shortener.png',
@@ -14,8 +13,7 @@ export const tools: Tool[] = [
   },
   {
     name: 'QR Generator',
-    description:
-      'Customize, brand, and share information.\nCraft QR codes for your unique needs.',
+    description: 'Customize, brand, and share information.\nCraft QR codes for your unique needs.',
     iconFilenames: ['qr.svg', 'qr_1.svg'],
     secondaryIconClass: 'right-[4px] top-0 w-[32%]',
     decorImgSrc: '/images/home/side_shortener.png',
@@ -26,8 +24,7 @@ export const tools: Tool[] = [
   },
   {
     name: 'GDSC Certificate',
-    description:
-      'Easily create, edit, export, and share certificates on our user-friendly platform.',
+    description: 'Easily create, edit, export, and share certificates on our user-friendly platform.',
     iconFilenames: ['cert.svg', 'cert_1.svg'],
     secondaryIconClass: 'top-[8px] right-[-8px] w-[32%]',
     decorImgSrc: '/images/home/side_shortener.png',
@@ -37,8 +34,7 @@ export const tools: Tool[] = [
   },
   {
     name: 'GDSC Calendar',
-    description:
-      'Effortlessly manage your calendar, create schedules, and oversee daily events.',
+    description: 'Effortlessly manage your calendar, create schedules, and oversee daily events.',
     iconFilenames: ['calendar.svg', 'calendar_1.svg'],
     secondaryIconClass: 'top-[4px] right-[-8px] w-[52%]',
     decorImgSrc: '/images/home/side_shortener.png',
@@ -48,8 +44,7 @@ export const tools: Tool[] = [
   },
   {
     name: 'Code with Me',
-    description:
-      'Code together in real time on our collaborative code-sharing website.',
+    description: 'Code together in real time on our collaborative code-sharing website.',
     iconFilenames: ['codewithme.svg', 'codewithme_1.svg'],
     secondaryIconClass: 'w-[28%] top-[4px] right-[-6px]',
     decorImgSrc: '/images/home/side_shortener.png',
@@ -59,8 +54,7 @@ export const tools: Tool[] = [
   },
   {
     name: 'GDSC Q&A',
-    description:
-      'Get answers and career advice from tech-savvy consultants or school experts.',
+    description: 'Get answers and career advice from tech-savvy consultants or school experts.',
     iconFilenames: ['qna.svg', 'qna_1.svg'],
     secondaryIconClass: 'top-[8px] right-[-4px] w-[32%]',
     decorImgSrc: '/images/home/side_shortener.png',
