+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './AuthFormField';
 export * from './AlertLevel';
+export * from './Tool';
